Move webdav methods to new structure
