--- app/webdav/server/index.js
@@ -1,7 +1,0 @@
-import './methods/addWebdavAccount';
-import './methods/removeWebdavAccount';
-import './methods/getWebdavFileList';
-import './methods/getWebdavFilePreview';
-import './methods/getFileFromWebdav';
-import './methods/uploadFileToWebdav';
-import './startup/settings';
