+++ server/api/methods/index.js
@@ -18,3 +18,10 @@ import './integrations/outgoing/replayOutgoingIntegration';
 import './integrations/outgoing/updateOutgoingIntegration';
 import './tokenpass/findTokenChannels';
 import './tokenpass/getChannelTokenpass';
+import './webdav/addWebdavAccount';
+import './webdav/getFileFromWebdav';
+import './webdav/getWebdavCredentials';
+import './webdav/getWebdavFileList';
+import './webdav/getWebdavFilePreview';
+import './webdav/removeWebdavAccount';
+import './webdav/uploadFileToWebdav';
