+++ server/api/methods/webdav/getWebdavCredentials.ts
@@ -1,4 +1,4 @@
-import { ServerCredentials } from '../lib/webdavClientAdapter';
+import { ServerCredentials } from './lib/webdavClientAdapter';
 
 export function getWebdavCredentials(account: ServerCredentials): ServerCredentials {
 	const cred = account.token ? { token: account.token } : {
