+++ server/importPackages.js
@@ -38,7 +38,6 @@ import '../app/ui-master/server';
 import '../app/ui-vrecord/server';
 import '../app/user-data-download';
 import '../app/videobridge/server';
-import '../app/webdav/server';
 import '../app/webrtc/server';
 import '../app/meteor-accounts-saml/server';
 import '../app/e2e/server';
